server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
psycopg[binary]>=3.2.0
python-json-logger>=2.0.8
pydantic-settings>=2.1.0
slowapi>=0.1.8
alembic>=1.13.0
pytest>=7.4.0
